Refactor: handle challenge name edited
