+++ src/pages/Home/Challenge/ManagePage/utils/handleChallengeName.ts
@@ -12,3 +12,4 @@ const handleChallengeName = <T>(
   };
 
 export default handleChallengeName;
+
